move builder outside
